refactor: extract dependencies module
